8-Bit Slots v1.1 Update
I had to get rid of my nested if statements and put "and" in between the
additional conditions so that when it checked for a special match and
didn't find one it wouldn't say the player lost if they had no
"Goombas". Example: Player gets two "Energys" and a "Sword". After going
through the special 3 matches and get to the check for the 2 pairs and 1
from the same game, it would see there are 2 "Energys". It would check
for the "MMLife" and seeing it wasn't there would not make "win" true
and not give the player their bet back. So all the nested if statements
had to change.

As well, on line 360, I had to change the "images" range to 8, so it
would show the eighth image, "Mario", on the reels.